Move around functions and files
